fix: TraitRepository key를 Integer -> Long으로 수정

+++ src/main/java/com/triple/backend/test/repository/TraitRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TraitRepository extends JpaRepository<Trait, Integer> {
+public interface TraitRepository extends JpaRepository<Trait, Long> {
 
     // 테스트의 성향 정보 조회
     List<Trait> findByTest(Test test);
